requirements.txt
flask==2.3.3
requests==2.31.0
python-dotenv==1.0.0
flask-cors==3.0.10
gunicorn==20.1.0